refactor(server): route builtin tools through filesystem/workspace-path
