+++ apps/server/src/tool/builtins/grep.ts
@@ -3,7 +3,7 @@ import type { Tool } from "@openomni/protocol";
 import { defineTool } from "../define";
 import { optionalBoolean, optionalString, requireString } from "../shared/input";
 import { errorResult, fromError, successResult } from "../shared/result";
-import { resolveContainedPath } from "../shared/workspace-path";
+import { resolveContainedPath } from "../../filesystem/workspace-path";
 import type { NativeTool } from "../types";
 import { GREP_PROMPT } from "./grep-prompt";
